pyproject.toml
[project]
name = "ecommerce-agent"
version = "0.1.0"
description = "AI Agent System"
readme = "README.md"
requires-python = ">=3.12,<3.14"
authors = [
    { name = "Dev", email = "wangweiqiang24@mails.ucas.edu.cn" }
]
dependencies = [
    "fastapi>=0.115.0,<0.120.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "asyncpg>=0.31.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "httpx>=0.28.1",
    "redis>=7.1.0",
    "langchain>=1.2.3",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.5",
    "alembic>=1.18.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=6.6.0",
    "typer<0.16.0",  # Prevent pulling annotated-doc typosquatting via huggingface-hub
    "tenacity>=9.1.2",
    "pyjwt>=2.10.1",
    "langgraph-checkpoint-redis>=0.3.2",
    "langsmith>=0.4.25",
    "celery>=5.6.2",
    "click-plugins==1.1.1",  # Pin to avoid malicious 1.1.1.2 version
    "websockets>=14.1",
    "bcrypt>=4.0.1,<5.0",
    "email-validator>=2.3.0",
    "qdrant-client>=1.16.0",
    "fastembed>=0.6.0",
    "slowapi>=0.1.9",
    "opentelemetry-instrumentation-fastapi>=0.49b",
    "opentelemetry-exporter-otlp>=1.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
# 确保 asyncio 插件使用 session 级别的 loop 作用域
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "ty>=0.0.29",
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["migrations/", "scripts/", "celery_worker.py"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501", "N802", "UP042", "B904", "E722", "SIM108", "SIM116", "B008"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty.environment]
python-version = "3.12"
